// src/api/generated/model/propertySchema.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { PropertySchemaOptions } from './propertySchemaOptions';

export interface PropertySchema {
  id?: string;
  name?: string;
  options?: PropertySchemaOptions;
  type?: string;
}