// src/api/generated/model/rowItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { RowItemContent } from './rowItemContent';
import type { RowItemProperties } from './rowItemProperties';
import type { Time } from './time';
import type { UserInfo } from './userInfo';

export interface RowItem {
  content?: RowItemContent;
  created_at?: Time;
  created_by?: string;
  created_by_user?: UserInfo;
  id?: string;
  properties?: RowItemProperties;
  show_in_sidebar?: boolean;
  updated_at?: Time;
  updated_by?: string;
  updated_by_user?: UserInfo;
}