// src/api/generated/model/getDatabaseResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { PropertySchema } from './propertySchema';
import type { Time } from './time';
import type { ViewConfig } from './viewConfig';

export interface GetDatabaseResponse {
  created_at?: Time;
  created_by?: string;
  default_view?: string;
  description?: string;
  document_id?: string;
  icon?: string;
  id?: string;
  name?: string;
  schema?: PropertySchema[];
  space_id?: string;
  type?: string;
  updated_at?: Time;
  views?: ViewConfig[];
}