// src/api/generated/model/getRowResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { GetRowResponseContent } from './getRowResponseContent';
import type { GetRowResponseProperties } from './getRowResponseProperties';
import type { Time } from './time';
import type { UserInfo } from './userInfo';

export interface GetRowResponse {
  content?: GetRowResponseContent;
  created_at?: Time;
  created_by?: string;
  created_by_user?: UserInfo;
  database_id?: string;
  id?: string;
  properties?: GetRowResponseProperties;
  show_in_sidebar?: boolean;
  updated_at?: Time;
  updated_by?: string;
  updated_by_user?: UserInfo;
}