// src/api/generated/model/createViewRequestType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */

export type CreateViewRequestType = typeof CreateViewRequestType[keyof typeof CreateViewRequestType];


export const CreateViewRequestType = {
  table: 'table',
  board: 'board',
  calendar: 'calendar',
  gallery: 'gallery',
  list: 'list',
  timeline: 'timeline',
} as const;